Remove redundant zeroing
std::vector::resize already zeroes the new samples, so we don't need
to use std::fill. See https://stackoverflow.com/a/73333009/10477326

However, we pass a second parameter to explicitly copy-init with zero
and to not default-init. Benchmarks showed the performance difference
between the two to be statistically insignificant, despite default-init
having memset.